+++ be/app/Http/Controllers/API/UserController.php
@@ -27,7 +27,7 @@ class UserController extends Controller
     public function store(Request $request)
     {
         $user = User::create($requests->all());
-        return response()->json($user, 201);
+        return response()->json($user, 200);
     }
 
     /**
@@ -52,8 +52,7 @@ class UserController extends Controller
     {
         $user->update($request->all());
         
-
-        return response()->json($user, 200);
+        return response()->json($user, 201);
     }
 
     public function delete(User $user)
