+++ be/app/Http/Controllers/AuthController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\StoreUserRequest;
+use App\Http\Requests\RegisterRequest;
 use App\Models\User;
 use App\Traits\HttpResponses;
 use Illuminate\Http\Request;
@@ -12,28 +12,19 @@ class AuthController extends Controller
 {
     use HttpResponses;
 
-    public function login()
-    {
-        return 'this is my login method';
-    }
-
-    public function register(StoreUserRequest $request)
+    public function register(RegisterRequest $request)
     {
         $request->validated($request->all());
 
         $user = User::create([
-            'fullname' => $request->fullname,
+            'name' => $request->name,
             'email' => $request->email,
             'password' => Hash::make($request->password),
         ]);
 
         return $this->success([
             'user' => $user,
-            'token' => $user->createToken('API Token of' . $user->fullname)->plainTextToken
+            'token' => $user->createToken('API Token of' . $user->name)->plainTextToken
         ]);
-    }
-    public function logout()
-    {
-        return response()->json('This is my logout method');
     }
 }
