+++ be/routes/api.php
@@ -22,7 +22,4 @@ use App\Http\Controllers\TasksController;
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
-Route::post('/login', [AuthController::class, 'login']);
 Route::post('/register', [AuthController::class, 'register']);
-Route::post('/logout', [AuthController::class, 'logout']);
-Route::resource('/tasks', TasksController::class);
