+++ fe/src/App.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
-import './App.css';
 import DashboardPage from './pages/user/dashboard/DashboardPage';
 import Register from './pages/user/register/Register';
+import './App.css';
 
 const App = () => (
   <Routes>
